requirements.txt
alpaca-py>=0.8.2
pandas>=1.5.0
numpy>=1.23.0
ta>=0.10.2
scikit-learn>=1.2.0
joblib>=1.2.0
rich>=13.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
PyYAML>=6.0
websockets>=10.0
requests>=2.28.0
pytz>=2023.3
matplotlib>=3.7.0
seaborn>=0.12.0
pyarrow>=11.0.0
